+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/service/notion/converter/contents/filter/BookmarkFilter.java
@@ -1,8 +1,8 @@
-package io.f12.notionlinkedblog.service.notion.blockconverter.filter;
+package io.f12.notionlinkedblog.service.notion.converter.contents.filter;
 
 import java.util.List;
 
-import io.f12.notionlinkedblog.service.notion.blockconverter.CheckAnnotations;
+import io.f12.notionlinkedblog.service.notion.converter.contents.CheckAnnotations;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import notion.api.v1.model.blocks.Block;
