+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/service/notion/converter/contents/filter/HeadingThreeFilter.java
@@ -1,8 +1,8 @@
-package io.f12.notionlinkedblog.service.notion.blockconverter.filter;
+package io.f12.notionlinkedblog.service.notion.converter.contents.filter;
 
 import java.util.List;
 
-import io.f12.notionlinkedblog.service.notion.blockconverter.CheckAnnotations;
+import io.f12.notionlinkedblog.service.notion.converter.contents.CheckAnnotations;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import notion.api.v1.model.blocks.Block;
@@ -10,20 +10,20 @@ import notion.api.v1.model.pages.PageProperty;
 
 @AllArgsConstructor
 @Builder
-public class HeadingOneFilter {
+public class HeadingThreeFilter {
 	private Block block;
 
 	public String doFilter() {
-		List<PageProperty.RichText> texts = block.asHeadingOne().getHeading1().getRichText();
+		List<PageProperty.RichText> texts = block.asHeadingThree().getHeading3().getRichText();
 		StringBuilder stringBuilder = new StringBuilder();
 		for (PageProperty.RichText text : texts) {
 			CheckAnnotations letterShape = new CheckAnnotations(text);
 			stringBuilder.append(letterShape.applyAnnotations(text));
 		}
-		return toHeadingOne(stringBuilder);
+		return toHeadingThree(stringBuilder);
 	}
 
-	private String toHeadingOne(StringBuilder sb) {
-		return "# " + sb.toString() + "\n";
+	private String toHeadingThree(StringBuilder sb) {
+		return "### " + sb.toString() + "\n";
 	}
 }
