+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/service/notion/converter/contents/filter/ParagraphFilter.java
@@ -1,8 +1,8 @@
-package io.f12.notionlinkedblog.service.notion.blockconverter.filter;
+package io.f12.notionlinkedblog.service.notion.converter.contents.filter;
 
 import java.util.List;
 
-import io.f12.notionlinkedblog.service.notion.blockconverter.CheckAnnotations;
+import io.f12.notionlinkedblog.service.notion.converter.contents.CheckAnnotations;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import notion.api.v1.model.blocks.Block;
@@ -10,18 +10,23 @@ import notion.api.v1.model.pages.PageProperty;
 
 @AllArgsConstructor
 @Builder
-public class BulletedListItemFilter {
+public class ParagraphFilter {
 	private Block block;
 
 	public String doFilter() {
-		List<PageProperty.RichText> texts = block.asBulletedListItem().getBulletedListItem().getRichText();
+		List<PageProperty.RichText> texts = block.asParagraph().getParagraph().getRichText();
 		StringBuilder stringBuilder = new StringBuilder();
 
 		for (PageProperty.RichText text : texts) {
 			CheckAnnotations letterShape = new CheckAnnotations(text);
 			stringBuilder.append(letterShape.applyAnnotations(text));
 		}
-		stringBuilder.append("\n");
-		return "- " + stringBuilder.toString();
+
+		stringBuilder.append("\n\n");
+		if (texts.size() == 0) {
+			stringBuilder.append("</br>").append("\n\n");
+		}
+		return stringBuilder.toString();
 	}
+
 }
