+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/service/notion/converter/contents/filter/QuoteFilter.java
@@ -1,8 +1,10 @@
-package io.f12.notionlinkedblog.service.notion.blockconverter.filter;
+package io.f12.notionlinkedblog.service.notion.converter.contents.filter;
 
 import java.util.List;
 
-import io.f12.notionlinkedblog.service.notion.blockconverter.CheckAnnotations;
+import org.springframework.util.StringUtils;
+
+import io.f12.notionlinkedblog.service.notion.converter.contents.CheckAnnotations;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import notion.api.v1.model.blocks.Block;
@@ -10,23 +12,19 @@ import notion.api.v1.model.pages.PageProperty;
 
 @AllArgsConstructor
 @Builder
-public class ToDoFilter {
+public class QuoteFilter {
 	private Block block;
 
 	public String doFilter() {
-		boolean isChecked = block.asToDo().getToDo().getChecked();
-		List<PageProperty.RichText> texts = block.asToDo().getToDo().getRichText();
+		List<PageProperty.RichText> texts = block.asQuote().getQuote().getRichText();
 		StringBuilder stringBuilder = new StringBuilder();
+		stringBuilder.append("> ");
 
 		for (PageProperty.RichText text : texts) {
 			CheckAnnotations letterShape = new CheckAnnotations(text);
-			if (isChecked) {
-				stringBuilder.append(" - [x] ");
-			} else {
-				stringBuilder.append(" - [ ] ");
-			}
 			stringBuilder.append(letterShape.applyAnnotations(text));
 		}
-		return stringBuilder.toString() + "\n";
+		String replace = StringUtils.replace(stringBuilder.toString(), "\n", "</br>");
+		return replace + "\n\n";
 	}
 }
