+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/service/notion/converter/contents/filter/ToDoFilter.java
@@ -1,10 +1,8 @@
-package io.f12.notionlinkedblog.service.notion.blockconverter.filter;
+package io.f12.notionlinkedblog.service.notion.converter.contents.filter;
 
 import java.util.List;
 
-import org.springframework.util.StringUtils;
-
-import io.f12.notionlinkedblog.service.notion.blockconverter.CheckAnnotations;
+import io.f12.notionlinkedblog.service.notion.converter.contents.CheckAnnotations;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import notion.api.v1.model.blocks.Block;
@@ -12,19 +10,23 @@ import notion.api.v1.model.pages.PageProperty;
 
 @AllArgsConstructor
 @Builder
-public class QuoteFilter {
+public class ToDoFilter {
 	private Block block;
 
 	public String doFilter() {
-		List<PageProperty.RichText> texts = block.asQuote().getQuote().getRichText();
+		boolean isChecked = block.asToDo().getToDo().getChecked();
+		List<PageProperty.RichText> texts = block.asToDo().getToDo().getRichText();
 		StringBuilder stringBuilder = new StringBuilder();
-		stringBuilder.append("> ");
 
 		for (PageProperty.RichText text : texts) {
 			CheckAnnotations letterShape = new CheckAnnotations(text);
+			if (isChecked) {
+				stringBuilder.append(" - [x] ");
+			} else {
+				stringBuilder.append(" - [ ] ");
+			}
 			stringBuilder.append(letterShape.applyAnnotations(text));
 		}
-		String replace = StringUtils.replace(stringBuilder.toString(), "\n", "</br>");
-		return replace + "\n\n";
+		return stringBuilder.toString() + "\n";
 	}
 }
