+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/service/notion/converter/contents/filter/ToggleBlockFilter.java
@@ -1,9 +1,9 @@
-package io.f12.notionlinkedblog.service.notion.blockconverter.filter;
+package io.f12.notionlinkedblog.service.notion.converter.contents.filter;
 
 import java.util.List;
 
-import io.f12.notionlinkedblog.service.notion.blockconverter.CheckAnnotations;
-import io.f12.notionlinkedblog.service.notion.blockconverter.NotionBlockConverter;
+import io.f12.notionlinkedblog.service.notion.converter.contents.CheckAnnotations;
+import io.f12.notionlinkedblog.service.notion.converter.contents.NotionBlockConverter;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import notion.api.v1.NotionClient;
